Test symbols declared in one file and used in another using dedicated unit tests
Filetest can parse the files it tests in any order, which breaks when there are
dependencies between files (first parse the file that uses a symbol, then check
in the file that declare it that it has been used). The two tests requiring this
strict ordering have been moved to dedicated unit tests.