Give code-completion items a match score
In expressions like "A:" (QML property binding), "A =", "A +", "A <=", etc, A
is used to give code-completion items a score based on their type. For instance,
items having the same type as A are given a 10 and are displayed in the
"best matches" category of code-completion items.

A side-effect of this commit is that CodeCompletionContext is now in a state
in which implementing function call-tips should be easy.