Export DBus interface for jumping to a specific line in a file
The currently exported DBus interface does not allow opening files & jumping to locations inside the file. This ability is useful for external tools, which could offer to switch to KDevelop with the cursor right at the problematic line of code.

So add optional line and column parameters to the "openDocumentSimple()" method.

In the future the kdevelop executable could be extended to support line number selection from the command line using this DBus call.

REVIEW: 118987