Set the prototype context of a function when it is assigned to an object member
"object.method = function(){}" sets the prototype of the anonymous function to
"object". If object is a function (that has a prototype), then the prototype
of the anonymous function is set to the prototype of object.

"var o={method: function(){}}" is also supported. Here, the prototype of the
method is set to "o".

Unit tests will come once "prototype" and "new" are supported.