Handle "this", "foo.prototype" and "new Class()" in ExpressionVisitor
These changes build on the previous commits and expose the function prototypes
to the user. This commit also adds a whole testsuite that asserts that these
features work:

(note: "prototype" in the code does not have the exact same definition as in
the JS language. It means "the context in which live the declarations that
will be visible in objects created by instantiating the function")

* "this" refers to the prototype of the current function
* "function Class(){}" has its own prototype
* The prototype of the anonymous function "Class.prototype.foo = function(){}"
  is the one of Class. This makes class methods possible.
* "Class.prototype.something" properly refers to "something" declared using
  "this.something = ...;" in Class()
* Instantiating a class yields an object in which every declaration of the class'
  prototype is visible. "var o = Class(); o.something" registers an use of
  "something" if it has been declared in Class.
* Everything continues to work if "prototype" is omitted. For instance,
  "var object = {method: function(){ this.b = 4; }};" declares an object having
  a field named "b".