Register uses of declarations in the proper contexts
Use ContextBuilder::contextOnNode to open and close contexts as needed, so that
UseBuilder can register the uses in the correct contexts. This commit also replaces
DUContext* with DUContextPointer in ParseSession so that deleted contexts don't crash
KDevelop.

As usual, when something is fixed in UseBuilder, some use counts have to change
in the unit tests. All the uses are now correct and no declaration gets counted
as an use.