Offer code-completion items for built-in Javascript types
Variables having type int, float, mixed (Object), function, bool and array are
now considered by the code-completion engine and UseBuilder as being instances
of Number, Object, Function, Boolean and Array. Properties of those classes can
be used with them.

A small quirk, though, that comes from the fact that FunctionType is not an
identified type and has no idea of what its declaration is:

function a() {}
var b = function() {}

a. // Displays the members of Object (and the ones of the prototype of a)
b. // Displays the members of Function (and does nothing with the prototype of b)

a is a FunctionDeclaration and its kind is Declaration::Type. b is a Declaration,
and its kind is Declaration::Instance. They therefore hit two different code paths
in Helper::getInternalContext, and those path could only be unified if FunctionType
becomes an IdentifiedType subclass.