Do not load the GitHub plugin when the Git plugin is not.
This is currently not enforceable via .desktop files and thus
this is only a workaround. But at least it ensures the
GitHub plugin is not loaded when the Git plugin is unavailable.

Note that we can still unload the git plugin at runtime which
would leave the GitHub plugin loaded and in an invalid state...