Make it possible to define plugin dependencies by name.
To do so, define a X-KDevelop-IRequired entry in the plugin's
.desktop file with the contents "interface@pluginname", e.g.:

X-KDevelop-IRequired=org.kdevelop.IBasicVersionControl@kdevgit

In this case, only the kdevgit plugin is checked for the given
IBasicVersionControl interface.

The advantage here is that this way, the depdendency is also properly
tracked when trying to unload a plugin.

CCMAIL: kdevelop-devel@kde.org
CCBUG: 336948