Retrieve out-of-line function definitions properly
This requires us to instantiate an Identifier with the
fully-qualified identifier of the function definition (i.e.
"SomeClass::someFunc"). This seems to be on par with how oldcpp handles
this. Just using "someFunc" as the identifier won't work in this case,
because the out-of-line definition doesn't have an enclosing namespace
that further qualifies the identifier.