Port over adapt-signature-assistant from oldcpp
We now have two differently implemented adapt signature assistants in
kdev-clang.
- Clang-based version (ClangAdaptSignatureAssistant)
  - Lacking some features of the oldcpp version (shortening typenames,
    ...)
- DUChain-based version imported from oldcpp

The DUChain-based version is now used, Clang-based one is still there to be
able to rip out code from it (there's some generally useful code in it)

The advantage of the DUChain-based version is that we don't need to
access data from Clang twice and that we can just use DUChain
infrastructure for finding the declaration and for generating the code
for the declaration/definition signature.
We don't need to hold a lock for Clang's AST either in this case.