Manually add imported parent context
DUContext::Import instances have to be customized by the plugin in order to avoid
the indirect resolution of imported contexts. Indirect context resolution consists
of storing the owner of the context in Import, then then using owner->internalContext
or owner->internalFunctionContext for function declarations. This last case causes
huge problems and must be avoided, hence the workaround.