Improve use-building timings
Please check: The biggest bottle-neck at the moment is
ClangHelpers::findDeclaration which in turn is calling
DUContext::findContextAt millions of times for the 'benchDUChainBuilder'
bench mark.

Avoid retrieving the declaration of a cursor by caching the mapping of
those.

For this particular benchmark, I get an improvment of around 15% in both
execution time and instruction cost on my machine.

REVIEW: 119539