New approach for the cmake plugin
So far we were executing the whole project to figure out the cmake
projects information. This worked really well on one hand, but then
it was high maintainance and error prone.
What this plugin does is to use the generated compile_commands.json
file to populate the include directories and defines, therefore
making the analysis unneeded.

That's not entirely true though, we'll still want to keep analyzing
the files to provide the rest of the information we have to offer
(and that currently is not available with the commtted version of
the plugin) but now we can do it after loading the project, without
rush and following the parsing information in KDevelop, treating
cmake more like a language additionally to as a project.

CCMAIL: kdevelop-devel@kde.org