More robust dependencies between files
Don't parse a file until all its dependencies (and transitive dependencies) are
parsed. This avoids bad imports and base classes, and also provides a substantial
speed boost (a file is parsed only two times: to discover its dependencies, and
when everything is okay).

This commit also ensures that Object does not inherit from Object (removes a
bunch of KDevPlatform warnings), and fixes tests/files/types.js so that an
auto-created declaration does not get tested when it shouldn't.