Better forward declaration support
Make sure we associate unresolved types with their forward declaration
(if possible). Introduce ClangHelpers::findForwardDeclaration for that.

There's no easy way to figure out the forward declaration of a type for
a given cursor with libclang, unfortunately.
For example, clang_getTypeDeclaration(..) will always point to the *definition*
when it gets a hold of it in the current TU.