Update DUChain data when the environment has changed.
This combines the include paths, defines and pch-path into a
hash which is stored on-disk and then later compared to the new
environment. If the hash differs, we trigger a reparse.

To prevent opened files from getting reparsed at startup when
no data from the project could be obtained, we add some more code
for this special purpose: We check whether we parsed before
with a known project and whether the new environment data also
comes from a project. If not then we rely only on the timestamp
of the file on whether to trigger a reparse or not. Otherwise, the
previous data (i.e. with known project) takes precedence.