Improve hashing of QualifiedIdentifiers.
Reuse KDevHash helper function and include m_explicitlyGlobal and
m_isExpression bit fields, as well as the size of identifiers in
the generation of the hash.

This should lead to better hashes, and esp. reduce conflicts for
QIDs with the same ID contents, but differing bit field values.

The static item repo version is increased to prevent loading
issues, as this changes the on-disk cache format.