Improve macro support
For uses inside macro expansions we now create uses with empty ranges.
This is similar to what oldcpp did (for a good reason). We don't want to
have uses inside macro expansion "overlap" the macro expansion itself.
At any rate, there's no good way to highlight the range of a use inside
a macro expansion anyway, since it doesn't have a physical
representation in the editor.