Cleanup code by getting rid of the "fast" aliases.
This code was just a horrible anti-pattern. Any decent compiler will
inline QVector usage and there is no need to keep the size in another
variable.

Furthermore note that QVector.resize will default-construct the
data members. This means 0 for numeric and pointer types.

Finally, we use int instead of unsigned (short) int to reference
the buckets now and generally cleanup the code here and there.