Fix more compile warnings and cleanup the code a bit more.
The warnings where about uint vs. int comparisons. The cleanup
removes m_bucketHashSize as its just an alias to the compile-time
const bucketHashSize. This also allows us to remove the runtime
allocation of m_firstBucketForHash and just make it a direct array
member variable. Note that the ItemRepository is usually constructed
on the heap so this is fine.