Optimize: Store a list of unserialized child contexts.
Profiling shows that the repeated calls to deserialize the contents
of the indexed child contexts is quite costly. We now add a cache
of the unserialized data, and the serialized contents only get
touched when we change the list. Otherwise, i.e. in the majority
of cases, we can directly operate on the cache which is quicker.

Profiling the code completion test case in kdev-qmljs shows:

before: 23,379,658,719 instructions ( +-  0.15% )
after: 21,957,230,364 instructions, ( +-  0.16% )