Optimize: Keep a cache of local declarations.
Similar to the cache of child contexts, this has a great impact
on the performance of the DUChain. Esp. findDeclarations is much
faster now as we don't have to repeatedly de-serialize the data.

The kdev-qmljs code completion unit test now runs approx. 13% faster.
The number of instructions also drops down from ca. 22 billion to
about 19.5 billion.