Reenable the app wizard and file template wizard
The idea of this patch is to reenable the app wizard and file template wizard of KDevelop in the framework branch. For this, I had to reactivate Grantlee and fix/port away a few other things:

- Reenable Grantlee (Unofficial version 1.0) and change some CMake files (See note)
- Port kdevfilters to the new Qt5 plugin system. The install location was fixed, so the plugin is correctly loaded by Grantlee
- Renable the app wizard and file template wizard plugin. Change the serching and write paths for templates
- Change the default paths of several models. Because the dirs() method of KComponentData is deprecated, I had to introdue methods for adding search paths to the model
- Fixed the unit tests (See testing comment)

Note: Grantlee provides a Qt5 branch called 'volatile_qt5_port. I will ask the current maintainer of Grantlee when this branch will result in a new (official) release

REVIEW: 119763