Bring back the WelcomePage, this time as a Document
After some discussion in the SDK sprint, we decided we want it to end up in
as a pinned tab. (and then decide if we like it)
So far, I turned it into a plugin so it can be properly loaded, probably
should move it to kdevplatform.
The Projects Dashboard should be included there.

A patch in Qt is still needed for it to work properly:
https://codereview.qt-project.org/#/c/92201/