Infrastructure for adding problems
This commit also adds problems when the plugin is absolutely sure that something
went wrong. Problems are added when a module a not been found, when the is a
mismatch between the type of a QML property and the type of the value assigned
to it (only when both types are sure), and a hint is given when an expression
given as parameter to a function does not match the parameter type.