Map indices to source/proxy model when activating/selecting items.
We have code in place to automagically add a proxy model for
filtering purposes in the OutputWidget. The client IOutputViewModel
does not know about that, and we must never query its API with
indices from the proxy model. Similarly, we cannot select/scroll
to indices from the output view model when we added the proxy
model inbetween.

This patch ensures we always properly map between the two models.

BUG: 338377