Do not update caches of system headers when the environment changed.
This greatly improves the performance when we edit two files with
different environments which each include some system includes.
Without this patch, we'd play cache ping-pong on the system includes.

NOTE: This is not yet optimal as we can still play cache ping-pong
for user headers as soon as the include paths or defines for a given
set of includers are different. But maybe this is a first step
and a good idea in general? Feedback welcome!

REVIEW: 119999