Don't use the session to check enabled state unless necessary
This behaves as a nearly pointless optimization. It's actually done
because we have at least one test that deletes the session, leading
to issues with CompilerManager which rightfully asserts that it's
loaded even after the session has expired.
It's not necessary to check the session for it, so we don't.