Do not crash when FileManagerListJob gets destroyed before finish.
It will the emit the finished signal via the KJob dtor at which
point the FMLJ data is not available anymore. To protect against
dangling pointers, we then iterate over the full list and ensure
the ptr is removed.

This is pretty ugly but I don't see a safe alternative.