Do not auto-start the FileManagerListJob.
This can lead to funny side-effects, esp. when we put such a job
into an ExecuteCompositeJob. Assume this:

ECJ:
  -> child1
  -> child2
  -> FMLJ

Now the ECJ gets started and previously we also started the FMLJ.
If the latter finishes before the previous jobs in the ECJ finished,
we could end up starting them *twice*...