Make PluginPage work correctly
Also work around KPageDialog emitting a signal where it should not:
Adding a page always causes currentPageChanged() to be emitted
and even worse it gets emitted with the before parameter set to null.
This makes no sense since the current page didn't actually change.

This caused a hard to debug crash, but at least it now makes sure that
the apply button is always correctly greyed out after pressing apply
even if the ConfigPage decides to emit a changed() signal while it runs
the apply() method