Reduce map look-ups
Don't check contains then fetch the value. Fetch the value and then make
sure this was found. In this case we can just use the value because we have
pointers. If the pointer is null is just like it wasn't initialized in the
first place.
Also prefer iterating using interators than through the keys and then
looking them up.