Drop util/ssh_client
This hasn't been compiled in ages, and I do not see the benefit of the tool
being available when there is a similar application in kdeutils:
kdessh.

kdessh is unmaintained right now, too, so apparently there isn't really a need for
the whole functionality.

PS: Saves me from porting this to KF5