Fix assertion about unknown projects in the IDAIM.
The problem was that projects opened before the IDAIM plugin gets
initialized where never added.

To add these in the ctor (synchronously! otherwise we get races),
we must restructure the code a bit. It is notoriously overdesigned
and really should be simplified. Why is the CompilerProvider a plugin?
Why is there ICompilerProvider? Why the distinction across so many
libs/plugins?

Well, more work for later.