KF5: Port KDialog -> QDialog
Some parts/plugins were left out, because I did not verify everything
was working after the porting scripts did their job
- plugins/cvs - (don't even know how to trigger the code path)
- plugins/subversion (lot's of dialogs involved)
- plugins/snippet/legacy/ (will be dropped anyway, not worth it)
- plugins/dashboard (not built, what's the state of that anyway?)

For the rest of dialogs, ported in this patch, I manually tested each
of them. So everything should still be working as expected.
Still putting this up for review to have another pair of eyes on the patch.

REVIEW: 121128