Only connect the lambda to BackgroundParser::parseJobFinished once.
Otherwise we'd connect many different lambdas to the signal whenever
a document gets activated. The addresses of these lambdas are also
different thus Qt::UniqueConnection won't work as intended here.

This should fix a big slow-down over time when using KF5 KDevelop!

To get safe access to the LanguageController via ICore::self(),
without tying ourselves to that in the public API, we delay the
creation of the StaticAssistantManager until the language controller
is initialized.