Start using the JSON file directly for the Git plugin.
This introduces some issues with translations, but for us currently
it is more important to get the correct list of interfaces provided
by the git plugin.

The CMake macro I added can be used in other places in the future to
do the same. But first we need to iron out more aspects of the JSON
support, I guess.