Further simplify handling of plugin JSON description files.
Since the plugins are now installed into a versioned _directory_,
i.e. .../lib64/plugins/kdevplatform/${PLUGIN_VERSION/kdevplugin.so,
we don't need another entry for the version in the JSON file itself.

This allows us to get rid of the configure_file step. The biggest
advantage, besides simplicity, is speed: When the JSON file is changed
we don't trigger a CMake configure re-run, which is relatively slow.