Fix porting bugs of KUrl/QUrl usage in OutputFilteringStrategy.
This is done by refactoring the code to use Path instead of QUrl.
The unit tests are also changed a bit to be more explicit. I hope
I didn't change the functionality in the process though.

Also note that ecm_mark_as_test is not working as intended. We'll
have to mark all tests using ecm_add_test instead.