Cleanup KDevHash and benchmark it against IndexedString::hashString.
Also cleanup the benchmark code a bit to ensure we actually benchmark
the correct thing without the compiler optimizing everything away.

RunningHash is still faster, but I have the gut feeling it produces
far worse hash values. I'll need to create other tests to verify this.
In the end, the higher cost of KDevHash might be worth it if it leads
to better behavior when used in a hash map.