Fix DUChain assertion madness when using static linking.
The DUChain code relied on static code initialization to register
items. This fails horribly when we link the DUChain library statically
into both, the code completion library as well as the plugin.

To overcome this limiation, we get rid of the REGISTER_DUCHAIN_ITEM
macro invocations, and expand the code of DUChainItemRegistrator
manually in a function that gets called when the plugin initializes.

Similarly, we unregister the types when the plugin is unloaded,
but really don't do that actually. Currently, that is not supported
at all, and probably also the reason why language plugins are never
unloaded... Something for the future to fix, probably by rewriting
parts of the DUChain.