Show per-project configuration pages from all plugins.
Do not filter based on their X-KDevelop-Category. This used to work
as the KCM and the actual plugin used to be two different plugins,
each with their own X-KDevelop-Category. With the new improved API
this is not required and we can just ask for the number of
per-project pages to do the filtering.

This fixes the regression which hid the custom defines-and-includes
config page from the per-project config dialog.