grepview: fix two crashes.
The first crash is a crash-on-exit, when the kill() method of a no
longer existing job is called. Guard againsts this crash by passing
the job to the connect statement, to disconnect the lambda when the
job gets destroyed.

The second crash happens when GrepOutputDelegate::sizeHint() is
called while using the combo-box of the grep output view to switch
to previous search results. Simply change the code to be more
defensive.

REVIEW: 121436