sublime: prevent accidentally adding nullptrs to ContainerPrivate::viewForWidget
The likely culprit here is using operator[] to access a map, which implicitly
creates an entry containing a null view. I have audited all other accesses of
viewForWidget with operator[] to verify that they happen in a context where
the entry is non-null (either because the access is used to assign a non-null
view pointer or because the access is guaranteed to crash or abort immediately
if it would create an incorrect nullptr entry).

BUG: 342037
REVIEW: 121600
(cherry picked from commit fa8ab02ebd8f88c7214538cdf4b7e8c1ac07cd98)