Move Q_DECLARE_OPERATORS_FOR_FLAGS out of namespaces
Having the macro expanded while not in the global namespace messes with the
precedence rules of operator| in a way that prevents the QFlags mechanism
from working correctly.

Observe one additional change which was necessary precisely because somebody
had previously used a C-style cast to work around this breakage.

REVIEW: 121626