FocusedTreeView: use a timer to batch recomputations of column width
We need to recompute the column width when a new row is inserted to ensure that
the horizontal scrollbar behaves properly.

Under certain circumstances (in particular: running an extremely spammy process
or filtering large outputs), rowInserted is called many times in quick succession,
which leads to essentially quadratic runtime behavior, because some of the
recomputation triggers a slow layout recomputation inside QTreeView. So we delay
the recomputation until we get back to the event loop.

Also re-use this timer for the auto-scroll behavior in the standard output view
(which is currently the only user of this class).

After this change, output views work more smoothly even on very large outputs.
Unfortunately, there seems to still be some quadratic runtime behavior in Qt itself.

REVIEW: 121880