Do not repeat TodoExtraction for included files.
The TodoExtractor currently can only tokenize the top level file
of a given TU. Thus, we repeat that for every included file
and associate the todos in the top file with the included files.

Once we figure out how to tokenize an included file, we can fix
this up to properly extract the TODOs from included files.