Do not apply header-check to folders or files without extension.
This fixes the unit test which also spotted that "iostream" was
not included in the code completion. The reason was that this
is not detected as a header by us, as it has no extension.

Furthermore, we also excluded subfolders here, as they are also not
matched by the isHeader check.

This regression was introduced by me in rev 721ef97d98c0cbacdb7.