Enable the DUChain imports cache for clang contexts.
Running duchainify -t 4 in the clang directory yields the following
performance numbers as obtained from perf stat -r 10:

Before:
       10.032869148 seconds time elapsed  +-  4.49% )
       1.256718903 seconds time elapsed ( +-  0.47% )

After:
       9.695671460 seconds time elapsed ( +-  3.90% )
       1.062835628 seconds time elapsed ( +-  0.63% )

Here, the high number is for running duchainify with CLEAR_DUCHAIN_DIR
environment variable set to 1. The smaller number is for running
duchainify once, and then doing the perf stat un the cached data.

So there is a small, but noticeable, gain in this benchmark. Should
be even more noticeable for larger projects.