Connect to KTextEditor::Document when this gets created.
Inside the DocumentController, no document is available as we
lazy-load this only when a view gets requested. Thus, the connection
to the reload signal was never created. With the recent change to
prevent urlChanged signals during reload, we ended up with reload
not re-triggering parse jobs in KDevelop. Funny, this was broken
for years but worked by chance :)