Ensure we don't assume that the TU-file is the file we are working on.
This is done by explicitly requesting a path to be passed to
ParseSession::file and adding ::mainFile with the old logic. This
probably will fix quite some potential issues, e.g. in the signature
assistant or code completion.

Furthermore, we should not triviall be able to share the parse session
between different contexts, if I'm not mistaken.

REVIEW: 122094