Optimize: Write defines to temporary file and pass that via -imacros.
Upstream clang is not efficient at handling large lists of CLI
arguments, which we used to pass in all the defines. This easily goes
into the hundreds of defines for the compiler-builtin stuff alone.

See also: http://llvm.org/bugs/show_bug.cgi?id=20999

Now, with this approach, the bottleneck is gone and kdev-clang is
much quicker!