Cleanup "Jump to Outputmark" implementation
This fixes several issues:
- grepview plugin depending on the standardoutputview plugin
- standardoutput providing a globally useful shortcut combo
- F4 / Shift+F4 jumping to two different documents in case there are
  output marks in both the output view and the grep view result view

This patch provides a new interface (IToolViewActionListener) which
can be implemented by plugins providing a tool view,
which in turn provides a list of output marks one could jump to. The
actual shortcuts got moved to kdevplatformshell instead.

The toolview "listening" to those shortcuts is selected on a
LRU basis -- The last "activated" tool view will receive the shortcut
events.

REVIEW: 122490