Always make bundled tarfiles contents root owned
Matching the owner in the tar files is with the owners on disk is great
for backups and others where the owner is important metadata.

For distribution of files though, it is just adding random bytes into
the tar files, so let's avoid that.

Also note that when untar'ing, unless you are root, files always ends up
being yours anyways.

REVIEW: 122665