Make it possible to extract groups and users from .reviewboardrc
Creates a new UpdateRequest job that will allow us to change settings in a
RR, given that it's not possible to pass the values when creating it (!).
I had to change how the jobs are structured, as this added a 3rd (and last)
kind of jsonrpc http call, which requires PUT.
The job will be spawned in parallel, won't hold the rest of the process
anyway.