Do not report subversion as being not found when it actually was
The feature summary requires a _FOUND variable that is named exactly as
the package that is being looked up. In this case KDevelop uses
SubversionLibrary but the variable it set was SUBVERSION_FOUND. Hence
feature_summary thought the package was not found and reported that even
if all subversion libraries/includes have been solved.

Conflicts:
	plugins/CMakeLists.txt