Never offer implementation offers for stuff of system headers.
This removes many many completions items that are not relevant.
In the process, the code completion shows up much quicker for me.

Note though that the implementation items are still shown for
functions of other files in a project, which might be unrelated.
This should be improved in the future.