Use buddy as fallback for headers without explicitly pinned TU.
In such cases, it's worthwile to use the source buddy as best-guess as
it allows us to get rid of the parse run for the "header TU" which
gets thrown away later on anyways.

Performance wise, this patch improves the index run on a 48 file
project using STL: It goes from ~17s down to ~14s, just b/c we safe
a few needless clang runs on "header TUs".

REVIEW: 122160