Do not look into the DUChain for the buddy-fallback in ClangIndex.
For this use-case the minimal buddy code is sufficient and we can
exclude the costly DUChain lookup. Furthermore, this work-arounds
a potential deadlock I encountered where the parse job asks the
DUChain for buddies and dead-locks with a background thread that
tries to access the background parser while holding a DUChain lock.

I.e.: Never lock the DUChain in ParseJob constructors.