Do not trigger excessive read/writes to symbol table on updates.
When the context of a declaration did not change, we do not need
to first remove it from the symbol table and put it in afterwards
again for no reason. This just wastes a lot of CPU for nothing.

This also showed a performance penalty introduced for a change
in kdev-php which should be cleaned up in the future. For now,
I only added a FIXME comment.