Remove obsolete code that was required for kdev-php.
Now that kdev-php is not abusing setInSymbolTable for a custom
feature anymore, we can cleanup the common code here and all
users of ClassDeclaration benefit from an improved performance.

Before, every call to add/remove baseClass would remove the item
from the symbol table and add it again, which is far from cheap.