Add outlineview plugin
This plugin provides an outline view for the current document. It is
very similar to the outline view from eclipse and should also work for
any language.

I initially wrote this plugin in 2010 and then forgot about it. I've now
ported it to KF5 and it seems to work without any major issues. The only
major thing that is missing is that it updates once the document has been
reparsed and not whenever it is saved.

REVIEW: 123562